Created Helper Class for PropertySheet's test Operation
Removed createNewDataOutput
Updated Abstract4diacUITests copyright header
