+++ tests/org.eclipse.fordiac.ide.test.ui/META-INF/MANIFEST.MF
@@ -9,6 +9,7 @@ Bundle-RequiredExecutionEnvironment: JavaSE-21
 Import-Package: org.eclipse.core.runtime,
  org.eclipse.draw2d.geometry,
  org.eclipse.ui.internal.views.properties.tabbed.view,
+ org.eclipse.ui.views.properties.tabbed,
  org.junit;version="4.13.2",
  org.junit.jupiter.api,
  org.junit.jupiter.api.extension,
@@ -21,7 +22,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.swtbot.swt.finder,
  org.eclipse.swtbot.eclipse.finder,
  org.eclipse.swtbot.eclipse.gef.finder,
- org.eclipse.swtbot.nebula.nattable.finder;bundle-version="4.2.1",
+ org.eclipse.swtbot.nebula.nattable.finder,
  org.eclipse.gef,
  org.eclipse.fordiac.ide.model.ui,
  org.eclipse.ui.forms,
@@ -35,4 +36,4 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.swt,
  org.hamcrest.core,
  org.eclipse.draw2d,
- org.hamcrest.library;bundle-version="[1.3.0,2.2.1)"
+ org.hamcrest.library
